Commit files generated by ShindigIntegrator

+++ modules/shindig_integrator/shindig/php/src/social/sample/ShindigIntegratorDbFetcher.php
@@ -1,5 +1,5 @@
 <?php
-// $Id: ShindigIntegratorDbFetcher.php,v 1.2.2.5 2009/08/13 09:42:55 impetus Exp $
+// $Id: ShindigIntegratorDbFetcher.php,v 1.2.2.11.2.1 2009/10/23 05:54:21 impetus Exp $
 /**
  * @file
  * DB interaction layer for shindig services
@@ -8,38 +8,95 @@
  * This module contains core shindig server
  */
 
-global $TEST_ENABLED;
-if (! $TEST_ENABLED ) {
-global $base_url;
-$base_url = ((isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] == 'on') ? 'https' : 'http') . '://'. $_SERVER['HTTP_HOST'];
-if($base_path = Config::get('drupal_base_path')) {
-  $base_url .= $base_path;
-}
-$dir = getcwd();
-chdir(Config::get('include_path').'/..');
-require_once './includes/bootstrap.inc';
-drupal_bootstrap(DRUPAL_BOOTSTRAP_FULL);
-chdir($dir);
-
-}
-
 class ShindigIntegratorDbFetcher {
 	private $db;
 	private $url_prefix;
-	
+ 	private $drupal_dir;
+
 	// Singleton
 	private static $fetcher;
 
 	private function __construct()
 	{
-		global $url_array;
-		$this->url_prefix = "http://" . $_SERVER['HTTP_HOST'] . $url_array['path']."/";
+    global $base_url;
+    //Initiliaze Drupal
+    if(!defined(DRUPAL_BOOTSTRAP_FULL)) {
+      $base_url = ((isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] == 'on') ? 'https' : 'http') . '://'. $_SERVER['HTTP_HOST'];
+      try {
+        $base_path = Config::get('drupal_base_path');
+        if($base_path) {
+          $base_url .= $base_path;
+        }
+      }
+      catch(Exception $exc) {
+        //Ignore config exception for drupal_base_path
+      }
+      $dir = getcwd();
+      try {
+        $this->drupal_dir = Config::get('drupal_dir');
+        chdir($this->drupal_dir);
+      }
+      catch(Exception $exc) {
+        //Walk the directory tree to Drupal root
+        //This doesn't work if this file realpath is not under the Drupal root
+        chdir(dirname(__FILE__));
+        $this->drupal_dir = getcwd();
+        chdir('..');
+        while(!file_exists('cron.php') && $this->drupal_dir != getcwd()) {
+          $this->drupal_dir = getcwd();
+          chdir('..');
+        }
+        $this->drupal_dir = getcwd();
+      }
+      require_once './includes/bootstrap.inc';
+      drupal_bootstrap(DRUPAL_BOOTSTRAP_FULL);
+      //Return to original dir
+      chdir($dir);
+    }
+		$this->url_prefix = $base_url;
 	}
 
 	private function __clone()
 	{
 		// private, don't allow cloning of a singleton
 	}
+
+  /**
+   * @see drupal_alter
+   */
+  private function drupalAlter($type, &$data) {
+    $dir = getcwd();
+    chdir($this->drupal_dir);
+
+    // Hang onto a reference to the data array so that it isn't blown away later.
+    $args = array(&$data);
+
+    // Now, use func_get_args() to pull in any additional parameters passed into
+    // the drupalAlter() call.
+    $additional_args = func_get_args();
+    array_shift($additional_args);
+    array_shift($additional_args);
+    $args = array_merge($args, $additional_args);
+
+    foreach (module_implements($type .'_alter') as $module) {
+      $function = $module .'_'. $type .'_alter';
+      call_user_func_array($function, $args);
+    }
+
+    chdir($dir);
+  }
+
+  /**
+   * @see module_invoke_all
+   */
+  private function drupalInvokeAll() {
+    $dir = getcwd();
+    chdir($this->drupal_dir);
+    $args = func_get_args();
+    $return = call_user_func_array('module_invoke_all', $args);
+    chdir($dir);
+    return $return;
+  }
 
 	static function get()
 	{
@@ -57,11 +114,11 @@ class ShindigIntegratorDbFetcher {
 	*   $userId for who friends are being retrived
 	* @return
 	*   $ret ids of all the friends of the user
-	*/	
+	*/
 	public function getFriendIds($user_id)
 	{
 		$ret = array();
-		$res = db_query("SELECT requester_id, requestee_id FROM user_relationships WHERE requester_id = %d OR requestee_id = %d AND rtid=(select rtid FROM user_relationship_types WHERE name='friend') AND approved=1", $user_id, $user_id);
+		$res = db_query("SELECT requester_id, requestee_id FROM {user_relationships} WHERE requester_id = %d OR requestee_id = %d AND rtid=(select rtid FROM {user_relationship_types} WHERE name='friend') AND approved=1", $user_id, $user_id);
 		while ($row = db_fetch_array($res)) {
 			if($row['requester_id']==$user_id)
 			{
@@ -74,6 +131,8 @@ class ShindigIntegratorDbFetcher {
                     $ret[] = $row['requester_id'] ;
 			}
 		}
+    //Invoke hook_friends_alter() implementations
+    $this->drupalAlter('friends', $ret, $user_id);
 		return $ret;
 	}    
 
@@ -96,47 +155,43 @@ class ShindigIntegratorDbFetcher {
 		foreach($ids as $id)
 		{
 			$user =array();
-			$res = db_query("SELECT * FROM {profile_values} INNER JOIN {profile_fields} ON {profile_values}.fid = {profile_fields}.fid WHERE uid =%d", $id);		
+			$res = db_query("SELECT * FROM {profile_values} INNER JOIN {profile_fields} ON {profile_values}.fid = {profile_fields}.fid WHERE uid =%d", $id);
 			if ($res) {
-				while ($row = db_fetch_array($res)) 
+				while ($row = db_fetch_array($res))
 				{
 					$user[$row['name']] = $row['value'];
 					$user['uid'] = $row['uid'];
-				}			
-			}	
-			$user_id = $user['uid'];
-			$name = new Name($user['profile_fname'] . ' ' . $user['profile_lname']);
-			$name->setGivenName($user['profile_fname']);
-			$name->setFamilyName($user['profile_lname']);
-			$address = new Address("UNSTRUCTUREDADDRESS");
-			$address->setLocality($user['profile_city']);
-			$address->setCountry($user['profile_country']);
-			$person = new Person($user['uid'], $name);
-			$person->setAddresses(array($address));
-			$person->setNickname($user['profile_nickname']);
-			$person->setProfileUrl($this->url_prefix . 'user/' . $user['uid']);
-			$person->setDisplayName($user['profile_fname'] . ' ' . $user['profile_lname']);
-			$res = db_query("SELECT picture FROM {users} WHERE uid = %d",$id);	
-			$row =  db_fetch_array($res);
-			$person->setThumbnailUrl(! empty($row['picture']) ? $this->url_prefix.$row['picture'] : '');
-			if ($user['profile_gender'] == 'Female') {
-			    $person->setGender('FEMALE');
-			} else {
-			    $person->setGender('MALE');
+				}
+				if(!empty($user)) {
+					$user_id = $user['uid'];
+					$name = new Name($user['profile_fname'] . ' ' . $user['profile_lname']);
+					$name->setGivenName($user['profile_fname']);
+					$name->setFamilyName($user['profile_lname']);
+					$address = new Address("UNSTRUCTUREDADDRESS");
+					$address->setLocality($user['profile_city']);
+					$address->setCountry($user['profile_country']);
+					$person = new Person($user['uid'], $name);
+					$person->setAddresses(array($address));
+					$person->setNickname($user['profile_nickname']);
+					$person->setProfileUrl($this->url_prefix . 'user/' . $user['uid']);
+					$person->setDisplayName($user['profile_fname'] . ' ' . $user['profile_lname']);
+					$res = db_query("SELECT picture FROM {users} WHERE uid = %d",$id);
+					$row =  db_fetch_array($res);
+					$person->setThumbnailUrl(! empty($row['picture']) ? $this->url_prefix.$row['picture'] : '');
+					if (strtolower($user['profile_gender']) == 'female') {
+						$person->setGender('FEMALE');
+					} else {
+						$person->setGender('MALE');
+					}
+					$ret[$user_id] = $person;
+				}
 			}
-			$ret[$user_id] = $person;
-			
 		}
-		//Chnage current dir to the Drupal directory, so drupal code is executed from
-		//the expected path.
-		$dir = getcwd();
-		chdir(Config::get('include_path').'/..');
-		drupal_alter('people', $ret);
-		//Restore original dir
-    chdir($dir);
+    //Invoke hook_people_alter() implementations
+		$this->drupalAlter('people', $ret, $profileDetails, $options);
 		return $ret;
 	}
-	
+
 	/**
 	* Get User Id(s) who have specific application
 	* Return array of Person Model Objects
@@ -148,7 +203,7 @@ class ShindigIntegratorDbFetcher {
 	*/
 	public function getPeopleWithApp($appId) {
 		$peopleWithApp = array();
-		$res = db_query("SELECT user_id FROM user_applications WHERE application_id=%d", $appId);
+		$res = db_query("SELECT user_id FROM {user_applications} WHERE application_id=%d", $appId);
 		while ($row = db_fetch_array($res)) {			
 				$peopleWithApp[] = $row['user_id'];
 		}
@@ -172,17 +227,23 @@ class ShindigIntegratorDbFetcher {
 		$user_id = $user_id;
 		$app_id = $app_id;
 		
-		if (empty($value)) {
-			if (! db_query("DELETE FROM {application_settings} WHERE application_id = %d AND user_id = %d AND name = '%s'", $appId, $userId, $key)) {
-				return false;
-			}
-		} else {
-			if (! db_query("INSERT INTO {application_settings} (application_id, user_id, name, value) VALUES (%d, %d, '%s', '%s') ON DUPLICATE KEY UPDATE value = '%s'", $appId, $userId, $key, $value, $value)) {
-				return false;
-			}
-		}
-		return true;
-	}
+    //Invoke hook_shindig_save_appdata() implementations
+    $modules_results = array_filter($this->drupalInvokeAll('shindig_save_appdata', $userId, $key, $value, $appId));
+    //If no implementation returned a TRUE value, save the appdata ourself
+    if(empty($modules_results)) {
+      if (empty($value)) {
+        if (! db_query("DELETE FROM {application_settings} WHERE application_id = %d AND user_id = %d AND name = '%s'", $appId, $userId, $key)) {
+          return false;
+        }
+      }
+      else {
+        if (! db_query("INSERT INTO {application_settings} (application_id, user_id, name, value) VALUES (%d, %d, '%s', '%s') ON DUPLICATE KEY UPDATE value = '%s'", $appId, $userId, $key, $value, $value)) {
+          return false;
+        }
+      }
+      return true;
+    }
+  }
     
 	/**
 	* To get PersonAppData
@@ -198,29 +259,38 @@ class ShindigIntegratorDbFetcher {
 	*/	
 	public function getAppData($ids, $keys, $app_id) {
     $data = array();
-    $placeholders_ids = array_fill(0, count($ids), "%d");
-    if (in_array("@all", $keys)) {
-      $res = db_query("SELECT user_id, name, value FROM application_settings WHERE application_id = %d AND user_id IN (" . implode(',', $placeholders_ids) . ")", $app_id, $ids);
-    } 
-	else {
-      $placeholders_keys = array_fill(0, count($keys), "%s");
-      $values = array();
-      $values[] = $app_id;
-      $values = array_merge($values, $ids);
-      $values = array_merge($values, $keys);
-      $res = db_query("SELECT user_id, name, value FROM application_settings WHERE application_id = %d 
+    if(count($ids)) {
+      $placeholders_ids = array_fill(0, count($ids), "%d");
+      if (in_array("@all", $keys)) {
+        $values = array();
+        $values[] = $app_id;
+        $values = array_merge($values, $ids);
+        $res = db_query("SELECT user_id, name, value FROM {application_settings} WHERE application_id = %d AND user_id IN (" . implode(',', $placeholders_ids) . ")", $values);
+      } 
+	  else {
+	    if(count($keys)) {
+        $placeholders_keys = array_fill(0, count($keys), "%s");
+        $values = array();
+        $values[] = $app_id;
+        $values = array_merge($values, $ids);
+        $values = array_merge($values, $keys);
+        $res = db_query("SELECT user_id, name, value FROM {application_settings} WHERE application_id = %d
                         AND user_id IN (" . implode(',', $placeholders_ids) . ") 
 						AND name IN ('". implode(',', $placeholders_keys )."')", $values);
-    }
-    while ($app_data = db_fetch_array($res)) {
-      $user_id = $app_data['user_id'];
-      if (! isset($user_id)) {
-        $data[$user_id] = array();
+	    }
       }
-      $key = $app_data['name'];
-      $value = $app_data['value'];
-      $data[$user_id][$key] = $value;
+      while ($app_data = db_fetch_array($res)) {
+        $user_id = $app_data['user_id'];
+        if (! isset($user_id)) {
+          $data[$user_id] = array();
+        }
+        $key = $app_data['name'];
+        $value = $app_data['value'];
+        $data[$user_id][$key] = $value;
+      }
     }
+    //Invoke hook_appdata_alter
+    $this->drupalAlter('appdata', $data, $app_id);
     return $data;
   }
     
@@ -238,9 +308,21 @@ class ShindigIntegratorDbFetcher {
 	{
 		$userId = $userId;
 		$appId = $appId;
-		if(!db_query("DELETE FROM {application_settings} WHERE application_id = %d AND user_id = %d AND name = '%s'", $appId, $userId, $key)) {
+
+    //Invoke hook_shindig_create_activites implementations
+    $modules_results = array_filter($this->drupalInvokeAll('shindig_delete_appdata',$appId, $userId, $key));
+    //If no implementation returned a TRUE value, delete the appdata ourself
+    if(empty($modules_results)) {
+		if($key == "*") {
+		  if(!db_query("DELETE FROM {application_settings} WHERE application_id = %d AND user_id = %d ", $appId, $userId)) {
 			return false;
+		  }
+		} else {
+		  if(!db_query("DELETE FROM {application_settings} WHERE application_id = %d AND user_id = %d AND name = '%s'", $appId, $userId, $key)) {
+			return false;
+		  }
 		}
+    }
 		return true;
 
 	}
@@ -260,10 +342,20 @@ class ShindigIntegratorDbFetcher {
 		$title = isset($activity['title']) ? trim($activity['title']) : '';
 		$body = isset($activity['body']) ? trim($activity['body']) : '';
 		$time = time();
-		db_query("INSERT INTO activities (id, user_id, app_id, title, body, created) VALUES (0, %d, %d, '%s', '%s', $time)", $user_id, $app_id, $title, $body);
-		if (! ($activityId = db_last_insert_id('activities', 'id'))) {
-			return false;
-		}
+    //Invoke hook_shindig_create_activites implementations
+    $modules_results = array_filter($this->drupalInvokeAll('shindig_create_activites', $user_id, $app_id, $title, $body));
+    //The activities has been saved, the first non-FALSE results is the activity ID
+    if(!empty($modules_results)) {
+      $activityId = reset($modules_results);
+    }
+    //If no implementation returned a TRUE value, save the activities ourself
+    else {
+      db_query("INSERT INTO {activities} (id, user_id, app_id, title, body, created) VALUES (0, %d, %d, '%s', '%s', $time)", $user_id, $app_id, $title, $body);
+      $activityId = db_last_insert_id('activities', 'id');
+    }
+    if (!$activityId) {
+      return false;
+    }
 		$mediaItems = isset($activity['mediaItems']) ? $activity['mediaItems'] :  array();
 		if (count($mediaItems)) {
 			foreach ($mediaItems as $mediaItem) {
@@ -273,7 +365,12 @@ class ShindigIntegratorDbFetcher {
 				$type = trim($type);
 				$mimeType = trim($mimeType);
 				$url = trim($url);
-				db_query("INSERT INTO activity_media_items (id, activity_id, mime_type, media_type, url) VALUES (0, %d, '%s', '%s', '%s')", $activityId, $mimeType, $type, $url);			
+        //Invoke hook_shindig_create_mediaitem implementations
+        $modules_results = array_filter($this->drupalInvokeAll('shindig_create_mediaitem', $activityId, $mimeType, $type, $url));
+        //If no implementation returned a TRUE value, save the mediaitem ourself
+        if(empty($modules_results)) {
+          db_query("INSERT INTO {activity_media_items} (id, activity_id, mime_type, media_type, url) VALUES (0, %d, '%s', '%s', '%s')", $activityId, $mimeType, $type, $url);
+        }
 			}
 		}
 		return true;
@@ -305,6 +402,9 @@ class ShindigIntegratorDbFetcher {
 		foreach ($ids as $key => $val) {
 			$ids[$key] = $val;
 		}
+  if(!is_array($ids) || count($ids) < 1) {
+    return $activities;
+  }
 		$res = db_query("
 			SELECT 
 				activities.user_id as user_id,
@@ -313,7 +413,7 @@ class ShindigIntegratorDbFetcher {
 				activities.body as activity_body,
 				activities.created as created
 			FROM 
-				activities
+				{activities}
 			WHERE
 				activities.user_id IN (" . implode(',', $ids) . ")
 			ORDER BY 
@@ -328,6 +428,8 @@ class ShindigIntegratorDbFetcher {
 			$activity->setMediaItems($this->getMediaItems($row['activity_id']));
 			$activities[] = $activity;
 		}
+    	//Invoke hook_activities_alter() implementations
+    	$this->drupalAlter('activities', $activities, $appId);
 		return $activities;
 	}
 
@@ -343,10 +445,12 @@ class ShindigIntegratorDbFetcher {
 	{
 		$media = array();
 		$activity_id = $activity_id;
-		$res = db_query("SELECT mime_type, media_type, url FROM activity_media_items WHERE activity_id = %d", $activity_id);
+		$res = db_query("SELECT mime_type, media_type, url FROM {activity_media_items} WHERE activity_id = %d", $activity_id);
 		while ($row = db_fetch_array($res)) {
-			$media[] = new MediaItem(strtoupper($row['mime_type']), strtoupper($row['media_type']), $row['url']);
+			$media[] = new MediaItem($row['mime_type'], strtoupper($row['media_type']), $row['url']);
 		}
+    	//Invoke hook_mediaitems_alter() implementations
+    	$this->drupalAlter('mediaitems', $media, $activity_id);
 		return $media;
 	}
